rename dir services/catalog-detail to catalogs
